some changes from the beginning of issues
