+++ code_measurement/project_euler/arithmetic/natural_numbers.py
